+++ src/Dmnk.Icons.Blazor.Fluent/FluentIconDefinition.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Components;
+﻿using Dmnk.Icons.Core;
+using Microsoft.AspNetCore.Components;
 using Fluent = Microsoft.FluentUI.AspNetCore.Components;
 
 namespace Dmnk.Icons.Blazor.Fluent;
@@ -7,9 +8,9 @@ internal class FluentIconDefinition(
     Fluent::Icon fluentIcon
 ) : CustomIconDefinition(fluentIcon.Name)
 {
-    public override MarkupString ToMarkup()
+    public override MarkupString ToMarkup(System.Drawing.Color? color = null)
     {
-        var markup = fluentIcon.ToMarkup().Value;
+        var markup = fluentIcon.ToMarkup(color: color?.ToHexString()).Value;
         // The Fluent ToMarkup() hard-codes "background-color: var(--neutral-layer-1);" into the SVG
         // style attribute.
         markup = markup.Replace("background-color: var(--neutral-layer-1); ", string.Empty);